Add lazy and basic wrappers for Secp256k1 curve, make original package-private
