+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add lazy and naive wrappers around `Secp256k1` curve, and make curve implementation package-private
+
 ## [2.1.0]
 
 ### Added
